Migrate table adding function to model.py
Start rewriting to avoid using dataframes after the initial load
